Avg emotion func, changed emotions from float to int

+++ src/GradReflect/NoteEntry+CoreDataProperties.swift
@@ -21,7 +21,7 @@ extension NoteEntry {
     @NSManaged public var situation: String
     @NSManaged public var thoughts: String
     //@NSManaged public var emotions: String //Float
-    @NSManaged public var emotionsScale: Float
+    @NSManaged public var emotionsScale: Int
     @NSManaged public var behaviour: String
     @NSManaged public var futureAlternate: String
     @NSManaged public var whyEmotions: String
